+++ packages/yoga/src/Divider/web/Divider.jsx
@@ -19,11 +19,9 @@ const StyledDivider = styled.hr`
   }}
 `;
 
-const Divider = props => {
-  const { vertical } = props;
-
-  return <StyledDivider vertical={vertical} />;
-};
+const Divider = React.forwardRef(({ vertical }, ref) => {
+  return <StyledDivider vertical={vertical} ref={ref} />;
+});
 
 Divider.propTypes = {
   /** If this value is defined, the divider will be in vertical if the flexDirection is row type */