+++ packages/yoga/src/Feedback/index.js
@@ -1,0 +1,3 @@
+import Feedback from './web';
+
+export default Feedback;