+++ packages/yoga/src/Feedback/web/index.js
@@ -1,0 +1,8 @@
+import Feedback from './Feedback';
+import { Title, PrimaryButton, SecondaryButton } from './StyledFeedback';
+
+Feedback.Title = Title;
+Feedback.PrimaryButton = PrimaryButton;
+Feedback.SecondaryButton = SecondaryButton;
+
+export default Feedback;