+++ packages/yoga/src/Heading/Heading.theme.js
@@ -1,0 +1,14 @@
+const Heading = ({ spacing, colors }) => ({
+  background: colors.white,
+  padding: {
+    vertical: spacing.small,
+    horizontal: spacing.huge,
+  },
+  height: 48,
+  button: {
+    width: 48,
+    height: spacing.xxlarge,
+  },
+});
+
+export default Heading;