+++ packages/yoga/src/Input/native/Tel.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import Input from './Input';
 
-const Tel = props => (
+const Tel = React.forwardRef((props, ref) => (
   <Input
+    ref={ref}
     {...props}
     keyboardType="phone-pad"
     autoCompleteType="tel"
     textContentType="telephoneNumber"
   />
-);
+));
 
 export default Tel;