+++ packages/yoga/src/index.native.js
@@ -1,4 +1,4 @@
-import ThemeProvider, { yogaTheme, theme } from './Theme';
+import ThemeProvider, { yogaTheme, theme, createTheme } from './Theme';
 import Button from './Button';
 import List from './List';
 import Checkbox from './Checkbox';
@@ -19,10 +19,13 @@ import Box from './Box';
 import Avatar from './Avatar';
 import Result from './Result';
 import Divider from './Divider';
+import Snackbar from './Snackbar';
+import Banner from './Banner';
 
 export {
   ThemeProvider,
   yogaTheme,
+  createTheme,
   theme,
   Button,
   List,
@@ -47,4 +50,6 @@ export {
   Avatar,
   Result,
   Divider,
+  Snackbar,
+  Banner,
 };