+++ packages/yoga/src/shared/propTypes/max.js
@@ -1,17 +1,19 @@
 import { checkPropTypes, number } from 'prop-types';
 
-const max = (value, proptypes = number) => (props, propName, componentName) => {
-  const { [propName]: valueProp } = props;
+const max =
+  (value, proptypes = number) =>
+  (props, propName, componentName) => {
+    const { [propName]: valueProp } = props;
 
-  checkPropTypes({ [propName]: proptypes }, props, 'prop', componentName);
+    checkPropTypes({ [propName]: proptypes }, props, 'prop', componentName);
 
-  if (valueProp > value) {
-    return new Error(
-      `Invalid prop ${propName} supplied to ${componentName}. ${propName} must be a value with maximum of ${value}`,
-    );
-  }
+    if (valueProp > value) {
+      return new Error(
+        `Invalid prop ${propName} supplied to ${componentName}. ${propName} must be a value with maximum of ${value}`,
+      );
+    }
 
-  return null;
-};
+    return null;
+  };
 
 export default max;